+++ modules/api_nelson/src/cpp/MatrixCheck.cpp
@@ -69,14 +69,6 @@ namespace Nelson {
         {
             return Aclass;
         }
-        if ((Aclass < NLS_INT64) || (Aclass == NLS_CHAR))
-        {
-            //  Aclass = NLS_INT64;
-        }
-        if ((Bclass < NLS_INT64) || (Bclass == NLS_CHAR))
-        {
-            //Bclass = NLS_INT64;
-        }
         // Division or matrix operations do no allow integer
         // data types.  These must be promoted to doubles.
         if (isDivOrMatrix && (Aclass < NLS_SINGLE))