+++ modules/interpreter/src/include/VertCatOperator.hpp
@@ -6,25 +6,23 @@
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation, either version 2 of the License, or
 // (at your option) any later version.
-// 
+//
 // This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
-// 
+//
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-function r = ndarraydouble_vertcat_double(varargin)
-	if (nargin() == 2)
-		if (all(size(varargin{2}) == [0 0]))
-			r = varargin{1};
-		else
-			error(_('Dimensions concatenated are not consistent.'));
-		end
-	else
-			error(_('Wrong number of input arguments.'));
-	end
-endfunction
+#pragma once
+//=============================================================================
+#include "nlsInterpreter_exports.h"
+#include "ArrayOf.hpp"
+#include "Evaluator.hpp"
+//=============================================================================
+namespace Nelson {
+    NLSINTERPRETER_IMPEXP ArrayOf VertCatOperator(Evaluator *eval, const ArrayOfVector &v);
+}
 //=============================================================================