+++ modules/single/builtin/cpp/Gateway.cpp
@@ -20,8 +20,6 @@
 #include "singleBuiltin.hpp"
 #include "single_dispBuiltin.hpp"
 #include "ndarraysingle_dispBuiltin.hpp"
-#include "single_vertcat_singleBuiltin.hpp"
-#include "ndarraysingle_vertcat_ndarraysingleBuiltin.hpp"
 #include "ndarraysingle_isequalBuiltin.hpp"
 #include "ndarraysingle_isequalnBuiltin.hpp"
 #include "single_isequalBuiltin.hpp"
@@ -38,8 +36,6 @@ static const nlsGateway gateway[] =
     { "single", Nelson::SingleGateway::singleBuiltin, 1, 1 },
     { "single_disp", Nelson::SingleGateway::single_dispBuiltin, 0, 1 },
     { "ndarraysingle_disp", Nelson::SingleGateway::ndarraysingle_dispBuiltin, 0, 1 },
-    { "single_vertcat_single", Nelson::SingleGateway::single_vertcat_singleBuiltin, 1, 2 },
-    { "ndarraysingle_vertcat_ndarraysingle", Nelson::SingleGateway::ndarraysingle_vertcat_ndarraysingleBuiltin, 1, 2 },
     { "ndarraysingle_isequal", Nelson::SingleGateway::ndarraysingle_isequalBuiltin, 1, 2 },
     { "ndarraysingle_isequaln", Nelson::SingleGateway::ndarraysingle_isequalnBuiltin, 1, 2 },
     { "single_isequal", Nelson::SingleGateway::single_isequalBuiltin, 1, 2 },